// src/app/api/product/product-opname/schema.ts
import { z } from 'zod';

const ProductSchema = z.object({
    id: z.number().optional(),
    product_id: z.number(),
    product_unit_id: z.number(),
    product_category_id: z.number().nullable().optional(),
    recorded_stock: z.number(),
    counted_stock: z.number(),
    difference_stock: z.number(),
    note: z.string().nullable().optional(),
});

export const formProductOpnameRequestSchema = z.object({
    location_id: z.string(),
    auto_approve: z.boolean().optional(),
    note: z.string().nullable().optional(),
    recorded_product_count: z.string().optional(),
    counted_product_count: z.string().optional(),
    difference_product_count: z.string().optional(),
    recorded_stock: z.string().optional(),
    counted_stock: z.string().optional(),
    difference_stock: z.string().optional(),
    products: z.array(ProductSchema),
});